# openapi.json
{
  "openapi": "3.0.2",
  "info": {
    "title": "TechFUL API",
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "http://127.0.0.1:8000",
      "description": "api"
    },
    {
      "url": "http://127.0.0.1:4010",
      "description": "mock"
    }
  ],
  "paths": {
...
    "/v1/jobs/associate_templates/{job_id}": {
      "put": {
        "tags": [
          "jobs"
        ],
        "summary": "スカウト求人テンプレート紐づけ更新",
        "description": "スカウト求人とテンプレートの紐づけを更新する",
        "operationId": "putJobsAssociateTemplates",
        "parameters": [
          {
            "description": "求人ID",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "integer",
              "description": "求人ID"
            },
            "name": "job_id",
            "in": "path"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/app__routers__v1__jobs__put_jobs_associate_templates__TemplateIds"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobsAssociateTemplateResponse"
                }
              }
            }
          },
          "422": {
            "description": "バリデーションエラー",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "jobs"
        ],
        "summary": "スカウト求人テンプレート紐づけ",
        "description": "スカウト求人をテンプレートに紐づけする",
        "operationId": "postJobsAssociateTemplates",
        "parameters": [
          {
            "description": "求人ID",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "integer",
              "description": "求人ID"
            },
            "name": "job_id",
            "in": "path"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/app__routers__v1__jobs__post_jobs_associate_templates__TemplateIds"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PostJobsAssociateTemplateResponse"
                }
              }
            }
          },
          "422": {
            "description": "バリデーションエラー",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/jobs/{job_id}": {
      "get": {
        "tags": [
          "jobs"
        ],
        "summary": "スカウト求人取得（特定）",
        "description": "特定のスカウト求人を取得する",
        "operationId": "getJobsJobId",
        "parameters": [
          {
            "description": "求人ID",
            "required": true,
            "schema": {
              "title": "Job Id",
              "minimum": 1.0,
              "type": "integer",
              "description": "求人ID"
            },
            "name": "job_id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/app__schemas__jobs__job_response__JobResponse"
                }
              }
            }
          },
          "422": {
            "description": "バリデーションエラー",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "jobs"
        ],
        "summary": "スカウト求人更新",
        "description": "スカウト求人を更新する",
        "operationId": "putJobs",
        "parameters": [
          {
            "description": "求人ID",
            "required": true,
            "schema": {
              "title": "Job Id",
              "minimum": 1.0,
              "type": "integer",
              "description": "求人ID"
            },
            "name": "job_id",
            "in": "path"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/app__routers__v1__jobs__put_jobs__Job"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/app__schemas__jobs__job_response__JobResponse"
                }
              }
            }
          },
          "422": {
            "description": "バリデーションエラー",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "jobs"
        ],
        "summary": "スカウト求人削除",
        "description": "スカウト求人を削除する",
        "operationId": "deleteJobsClose",
        "parameters": [
          {
            "description": "求人ID",
            "required": true,
            "schema": {
              "title": "Job Id",
              "minimum": 1.0,
              "type": "integer",
              "description": "求人ID"
            },
            "name": "job_id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "バリデーションエラー",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "JobsAssociateTemplateResponse": {
        "title": "JobsAssociateTemplateResponse",
        "type": "object",
        "properties": {
          "templateData": {
            "title": "Templatedata",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/app__schemas__jobs__jobs_associate_template_response__TemplateData"
            },
            "description": "テンプレート情報",
            "nullable": true
          }
        }
      },
...
      "PostJobsAssociateTemplateResponse": {
        "title": "PostJobsAssociateTemplateResponse",
        "type": "object",
        "properties": {
          "templateData": {
            "title": "Templatedata",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/app__schemas__jobs__post_jobs_associate_template_response__TemplateData"
            },
            "description": "テンプレート情報",
            "nullable": true
          }
        }
      },
...
      "ValidationError": {
        "title": "ValidationError",
        "required": [
          "errorCode",
          "message",
          "detail"
        ],
        "type": "object",
        "properties": {
          "errorCode": {
            "title": "Errorcode",
            "type": "string",
            "example": "VALIDATION_ERROR"
          },
          "message": {
            "title": "Message",
            "type": "string",
            "example": "入力値が不正です"
          },
          "detail": {
            "title": "Detail",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ValidationErrorDetail"
            }
          }
        }
      },
      "ValidationErrorDetail": {
        "title": "ValidationErrorDetail",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "type": "object",
        "properties": {
          "loc": {
            "title": "Loc",
            "type": "array",
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            }
          },
          "msg": {
            "title": "Msg",
            "type": "string"
          },
          "type": {
            "title": "Type",
            "type": "string"
          }
        }
      },
...
      "app__routers__v1__jobs__post_jobs_associate_templates__TemplateIds": {
        "title": "TemplateIds",
        "required": [
          "templateIds"
        ],
        "type": "object",
        "properties": {
          "templateIds": {
            "title": "Templateids",
            "type": "array",
            "items": {
              "type": "integer"
            },
            "description": "テンプレートID"
          }
        }
      },
...
      "app__routers__v1__jobs__put_jobs_associate_templates__TemplateIds": {
        "title": "TemplateIds",
        "required": [
          "templateIds"
        ],
        "type": "object",
        "properties": {
          "templateIds": {
            "title": "Templateids",
            "type": "array",
            "items": {
              "type": "integer"
            },
            "description": "テンプレートID"
          }
        }
      },
...
      "app__schemas__jobs__job_response__Areas": {
        "title": "Areas",
        "required": [
          "areaId",
          "areaName"
        ],
        "type": "object",
        "properties": {
          "areaId": {
            "title": "Areaid",
            "minimum": 1.0,
            "type": "integer",
            "description": "エリアID"
          },
          "areaName": {
            "title": "Areaname",
            "type": "string",
            "description": "エリア名"
          }
        }
      },
...
      "app__schemas__jobs__job_response__Features": {
        "title": "Features",
        "required": [
          "featureId",
          "featureName"
        ],
        "type": "object",
        "properties": {
          "featureId": {
            "title": "Featureid",
            "minimum": 1.0,
            "type": "integer",
            "description": "特徴ID"
          },
          "featureName": {
            "title": "Featurename",
            "type": "string",
            "description": "特徴名"
          }
        }
      },
      "app__schemas__jobs__job_response__Frameworks": {
        "title": "Frameworks",
        "required": [
          "frameworkId",
          "frameworkName"
        ],
        "type": "object",
        "properties": {
          "frameworkId": {
            "title": "Frameworkid",
            "minimum": 1.0,
            "type": "integer",
            "description": "フレームワークID"
          },
          "frameworkName": {
            "title": "Frameworkname",
            "type": "string",
            "description": "フレームワーク名"
          }
        }
      },
      "app__schemas__jobs__job_response__JobResponse": {
        "title": "JobResponse",
        "required": [
          "jobTitle",
          "updateAt",
          "occupation",
          "jobDescription",
          "workLocation",
          "initiativesToPreventPassiveSmoking",
          "employmentStatus",
          "contractPeriod",
          "trialPeriod",
          "socialInsurance",
          "occupationalAbility"
        ],
        "type": "object",
        "properties": {
          "jobTitle": {
            "title": "Jobtitle",
            "type": "string",
            "description": "求人ポジション名"
          },
          "skillLevel": {
            "title": "Skilllevel",
            "minimum": 0.0,
            "type": "integer",
            "description": "スキルレベル",
            "nullable": true
          },
          "jobImage": {
            "title": "Jobimage",
            "type": "string",
            "description": "求人画像",
            "nullable": true
          },
          "salaryForm": {
            "title": "Salaryform",
            "minimum": 1.0,
            "type": "integer",
            "description": "給与形態",
            "nullable": true
          },
          "minSalaryAmount": {
            "title": "Minsalaryamount",
            "type": "integer",
            "description": "最小給与額",
            "nullable": true
          },
          "maxSalaryAmount": {
            "title": "Maxsalaryamount",
            "type": "integer",
            "description": "最大給与額",
            "nullable": true
          },
          "areas": {
            "title": "Areas",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/app__schemas__jobs__job_response__Areas"
            },
            "description": "エリア",
            "nullable": true
          },
          "features": {
            "title": "Features",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/app__schemas__jobs__job_response__Features"
            },
            "description": "特徴",
            "nullable": true
          },
          "programmingLanguages": {
            "title": "Programminglanguages",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/app__schemas__jobs__job_response__ProgrammingLanguages"
            },
            "description": "プログラミング言語",
            "nullable": true
          },
          "frameworks": {
            "title": "Frameworks",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/app__schemas__jobs__job_response__Frameworks"
            },
            "description": "フレームワーク",
            "nullable": true
          },
          "skills": {
            "title": "Skills",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/app__schemas__jobs__job_response__Skills"
            },
            "description": "スキル",
            "nullable": true
          },
          "updateAt": {
            "title": "Updateat",
            "type": "string",
            "description": "最終更新日",
            "format": "date"
          },
          "publicationStartDate": {
            "title": "Publicationstartdate",
            "type": "string",
            "description": "募集開始日",
            "format": "date",
            "nullable": true
          },
          "publicationEndDate": {
            "title": "Publicationenddate",
            "type": "string",
            "description": "募集終了日",
            "format": "date",
            "nullable": true
          },
          "occupation": {
            "title": "Occupation",
            "allOf": [
              {
                "$ref": "#/components/schemas/app__schemas__jobs__job_response__Occupation"
              }
            ],
            "description": "職種"
          },
          "jobDescription": {
            "title": "Jobdescription",
            "type": "string",
            "description": "仕事内容"
          },
          "targetPeople": {
            "title": "Targetpeople",
            "type": "string",
            "description": "対象となる方",
            "nullable": true
          },
          "workLocation": {
            "title": "Worklocation",
            "type": "string",
            "description": "勤務地"
          },
          "numberOfPeopleHired": {
            "title": "Numberofpeoplehired",
            "type": "integer",
            "description": "採用人数",
            "nullable": true
          },
          "initiativesToPreventPassiveSmoking": {
            "title": "Initiativestopreventpassivesmoking",
            "type": "integer",
            "description": "集合場所における受動喫煙防止の取り組み"
          },
          "employmentStatus": {
            "title": "Employmentstatus",
            "type": "integer",
            "description": "雇用形態"
          },
          "contractPeriod": {
            "title": "Contractperiod",
            "type": "string",
            "description": "契約期間"
          },
          "trialPeriod": {
            "title": "Trialperiod",
            "type": "integer",
            "description": "試用期間"
          },
          "salarySupplement": {
            "title": "Salarysupplement",
            "type": "string",
            "description": "給与補足",
            "nullable": true
          },
          "salaryIncrement": {
            "title": "Salaryincrement",
            "type": "string",
            "description": "昇給",
            "nullable": true
          },
          "bonus": {
            "title": "Bonus",
            "type": "string",
            "description": "賞与",
            "nullable": true
          },
          "commutingTime": {
            "title": "Commutingtime",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/app__schemas__jobs__job_response__CommutingTime"
            },
            "description": "勤務時間",
            "nullable": true
          },
          "commutingTimeSupplement": {
            "title": "Commutingtimesupplement",
            "type": "string",
            "description": "勤務時間補足",
            "nullable": true
          },
          "overtimeWork": {
            "title": "Overtimework",
            "type": "boolean",
            "description": "時間外勤務",
            "nullable": true
          },
          "overtimeWorkFreeInput": {
            "title": "Overtimeworkfreeinput",
            "type": "string",
            "description": "時間外勤務の自由入力",
            "nullable": true
          },
          "holidaysOrVacation": {
            "title": "Holidaysorvacation",
            "type": "string",
            "description": "休日・休暇",
            "nullable": true
          },
          "socialInsurance": {
            "title": "Socialinsurance",
            "type": "string",
            "description": "社会保険"
          },
          "welfare": {
            "title": "Welfare",
            "type": "string",
            "description": "福利厚生",
            "nullable": true
          },
          "trainingSystem": {
            "title": "Trainingsystem",
            "type": "string",
            "description": "研修制度",
            "nullable": true
          },
          "recruitmentFlows": {
            "title": "Recruitmentflows",
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "採用フロー",
            "nullable": true
          },
          "selectionMethod": {
            "title": "Selectionmethod",
            "type": "string",
            "description": "選考方法",
            "nullable": true
          },
          "submittedDocument": {
            "title": "Submitteddocument",
            "type": "string",
            "description": "提出資料",
            "nullable": true
          },
          "recruitment": {
            "title": "Recruitment",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/app__schemas__jobs__job_response__Recruitment"
            },
            "description": "募集採用に関する状況に関して",
            "nullable": true
          },
          "occupationalAbility": {
            "title": "Occupationalability",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/app__schemas__jobs__job_response__OccupationalAbility"
            },
            "description": "職業能力の開発・向上に関する状況に関して"
          },
          "employmentManagement": {
            "title": "Employmentmanagement",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/app__schemas__jobs__job_response__EmploymentManagement"
            },
            "description": "企業における雇用管理に関して",
            "nullable": true
          },
          "contactInformation": {
            "title": "Contactinformation",
            "type": "string",
            "description": "\t問い合わせ先",
            "nullable": true
          },
          "corporatePr": {
            "title": "Corporatepr",
            "type": "string",
            "description": "企業PR",
            "nullable": true
          },
          "originalContent1Title": {
            "title": "Originalcontent1Title",
            "type": "string",
            "description": "オリジナルコンテンツ1タイトル",
            "nullable": true
          },
          "originalContent1Text": {
            "title": "Originalcontent1Text",
            "type": "string",
            "description": "オリジナルコンテンツ1本文",
            "nullable": true
          },
          "originalContent2Title": {
            "title": "Originalcontent2Title",
            "type": "string",
            "description": "オリジナルコンテンツ2タイトル",
            "nullable": true
          },
          "originalContent2Text": {
            "title": "Originalcontent2Text",
            "type": "string",
            "description": "オリジナルコンテンツ2本文",
            "nullable": true
          }
        }
      },
...
      "app__schemas__jobs__jobs_associate_template_response__TemplateData": {
        "title": "TemplateData",
        "required": [
          "templateId",
          "templateName"
        ],
        "type": "object",
        "properties": {
          "templateId": {
            "title": "Templateid",
            "type": "integer",
            "description": "テンプレートID"
          },
          "templateName": {
            "title": "Templatename",
            "type": "string",
            "description": "テンプレート名"
          }
        }
      },
...
      "app__schemas__jobs__post_jobs_associate_template_response__TemplateData": {
        "title": "TemplateData",
        "required": [
          "templateId",
          "templateName"
        ],
        "type": "object",
        "properties": {
          "templateId": {
            "title": "Templateid",
            "type": "integer",
            "description": "テンプレートID"
          },
          "templateName": {
            "title": "Templatename",
            "type": "string",
            "description": "テンプレート名"
          }
        }
      },
...
    }
  }
}